+++ pkg/application/applyBlueprintSpecUseCase_test.go
@@ -361,7 +361,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthUpfront(t *testing.T) {
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(&domain.BlueprintSpec{}, nil)
 
 		healthMock := newMockEcosystemHealthUseCase(t)
-		healthMock.EXPECT().WaitForHealthyEcosystem(testCtx, false, false).Return(ecosystem.HealthResult{}, assert.AnError)
+		healthMock.EXPECT().CheckEcosystemHealth(testCtx, false, false).Return(ecosystem.HealthResult{}, assert.AnError)
 
 		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
 
@@ -381,7 +381,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthUpfront(t *testing.T) {
 		repoMock.EXPECT().Update(testCtx, blueprintSpec).Return(assert.AnError)
 
 		healthMock := newMockEcosystemHealthUseCase(t)
-		healthMock.EXPECT().WaitForHealthyEcosystem(mock.Anything, false, false).Return(ecosystem.HealthResult{}, nil)
+		healthMock.EXPECT().CheckEcosystemHealth(mock.Anything, false, false).Return(ecosystem.HealthResult{}, nil)
 
 		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
 
@@ -404,7 +404,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthUpfront(t *testing.T) {
 		repoMock.EXPECT().Update(testCtx, blueprintSpec).Return(nil)
 
 		healthMock := newMockEcosystemHealthUseCase(t)
-		healthMock.EXPECT().WaitForHealthyEcosystem(mock.Anything, true, true).Return(ecosystem.HealthResult{}, nil)
+		healthMock.EXPECT().CheckEcosystemHealth(mock.Anything, true, true).Return(ecosystem.HealthResult{}, nil)
 
 		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
 
@@ -422,7 +422,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthUpfront(t *testing.T) {
 		repoMock.EXPECT().Update(testCtx, blueprintSpec).Return(nil)
 
 		healthMock := newMockEcosystemHealthUseCase(t)
-		healthMock.EXPECT().WaitForHealthyEcosystem(mock.Anything, false, false).Return(ecosystem.HealthResult{}, nil)
+		healthMock.EXPECT().CheckEcosystemHealth(mock.Anything, false, false).Return(ecosystem.HealthResult{}, nil)
 
 		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
 
@@ -457,7 +457,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthAfterwards(t *testing.T) 
 		repoMock.EXPECT().GetById(testCtx, blueprintId).Return(&domain.BlueprintSpec{}, nil)
 
 		healthMock := newMockEcosystemHealthUseCase(t)
-		healthMock.EXPECT().WaitForHealthyEcosystem(testCtx, false, false).Return(ecosystem.HealthResult{}, assert.AnError)
+		healthMock.EXPECT().CheckEcosystemHealth(testCtx, false, false).Return(ecosystem.HealthResult{}, assert.AnError)
 
 		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
 
@@ -478,7 +478,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthAfterwards(t *testing.T) 
 		repoMock.EXPECT().Update(testCtx, blueprintSpec).Return(assert.AnError)
 
 		healthMock := newMockEcosystemHealthUseCase(t)
-		healthMock.EXPECT().WaitForHealthyEcosystem(testCtx, false, false).Return(ecosystem.HealthResult{}, nil)
+		healthMock.EXPECT().CheckEcosystemHealth(testCtx, false, false).Return(ecosystem.HealthResult{}, nil)
 
 		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)
 
@@ -499,7 +499,7 @@ func TestApplyBlueprintSpecUseCase_CheckEcosystemHealthAfterwards(t *testing.T) 
 		repoMock.EXPECT().Update(testCtx, blueprintSpec).Return(nil)
 
 		healthMock := newMockEcosystemHealthUseCase(t)
-		healthMock.EXPECT().WaitForHealthyEcosystem(testCtx, false, false).Return(ecosystem.HealthResult{}, nil)
+		healthMock.EXPECT().CheckEcosystemHealth(testCtx, false, false).Return(ecosystem.HealthResult{}, nil)
 
 		sut := NewApplyBlueprintSpecUseCase(repoMock, nil, healthMock, nil)